+++ whatshap/cli/phase.py
@@ -53,7 +53,7 @@ from whatshap.pedigree import (
     Trio,
 )
 from whatshap.timer import StageTimer
-from whatshap.utils import ChromosomeFilter, plural_s, warn_once
+from whatshap.utils import ChromosomeFilter, plural_s, raise_if_any_sample_not_in_vcf, warn_once
 from whatshap.cli import CommandLineError, log_memory_usage, PhasedInputReader
 from whatshap.merge import ReadMerger, DoNothingReadMerger, ReadMergerBase
 from whatshap.types import PhasingAlgorithm
@@ -735,13 +735,6 @@ def log_best_case_phasing_info(readset: ReadSet, selected_reads: ReadSet) -> Non
         "... would be %d non-singleton phased blocks without read selection",
         n_best_case_nonsingleton_blocks,
     )
-
-
-def raise_if_any_sample_not_in_vcf(vcf_reader: VcfReader, samples: Sequence[str]) -> None:
-    vcf_sample_set = set(vcf_reader.samples)
-    for sample in samples:
-        if sample not in vcf_sample_set:
-            raise CommandLineError(f"Sample {sample!r} requested on command-line not found in VCF")
 
 
 def setup_families(
